Workable AlexNet model.
We freeze all the convolutional layers but leave the fully connected ones able to learn at their normal rate. For our new final layer, we set the learning rate multiplier to be 10 for normal weights and 20 for biases.

Stepsize is 1/4th the normal rate, while the base learning rate is 0.001 (as opposed to AlexNet’s normal 0.01 rate).