Move bounded validation images into their own directory.
Prepares the way for data augmentation - we only want to do this on training images, not validation ones.

Also brings clarity to when we generate test bounding box images — we only want to do those on validation images, not training ones, as the trained network could overfit to training data.